hermes: Update Lpd::.* function parameters
Previously, the base::Closure supplied by the user was both saved in a
private member variable |user_success_| and also passed as a parameter
to every callback we registered with the Esim interface. This didn't
make a lot of sense, so I removed the function parameter and kept the
private member variable.

BUG=b:80591752
BUG=chromium:847619
TEST=emerge-cheza hermes

Change-Id: I22944b8894213823de51dc760bfa5cc4cf70b3cb
Reviewed-on: https://chromium-review.googlesource.com/1171346
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>