arc-setup: Make GenerateHostSideCode() failures fatal
CL:1166448 has removed the only known issue that causes
GenerateHostSideCode() to fail. This CL wraps the function call
with EXIT_IF to turn down the mini container when the relocation
function fails.

This should be better because without the relocated boot files,
the container won't fully work anyway, and allowing the system to
continue to run such a half-working container is making it harder
to debug auto-test flakiness and as such.

BUG=chromium:871227
TEST=cheets_ContainerSmokeTest

Change-Id: Ie7534f6b022162f985c59dad3567b17187dd5ac5
Reviewed-on: https://chromium-review.googlesource.com/1171242
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Xiaoyong Zhou <xzhou@chromium.org>