bluetooth: Handle Pair and CancelPairing methods for org.bluez.Device1 interface
This adds a helper function for registering an async D-Bus handler to object
manager wrapper, so that we can register the handlers of Pair and CancelPairing
and reply to the method call later. This also adds unit tests for Pair
and CancelPair to newblue_unittest.

There can be only one active pairing at a time, and this is enforced by
having a struct representing the ongoing pairing session. Once newblued hears
back from pair state changes, we send the response(s) to the method call(s).

Based on the appearance/type of the devices, newblued set the security
requirement accordingly. For instance, MITM protection is set to false
for mouses but true for keyboards.

It is worth noting that security manager's IO capability is set to "no
input no display" in the first place until any client register itself as
the observer of passkey display/request; therefore this patch is tested
only with BLE mouses.

BUG=chromium:863613
TEST=(1) run FEATURES=test emerge-$BOARD bluetooth and verify that all unit
         tests pass
     (2) run newblued manually with
         restart bluetoothd && minijail0 -u bluetooth -- /usr/bin/newblued
         and verify that Pair/CancelPairing method calls work with
         bluetoothctl --service=org.chromium.Newblue
CQ-DEPEND=CL:1186082

Change-Id: Icc66276c6c998e27d131a0d694e87e176608a497
Reviewed-on: https://chromium-review.googlesource.com/1187568
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>