portier: Proxy Group class.
This change introduces a Proxy Group class which can contain / manage
a set of group members (known as ProxyGroupMember).  The ProxyGroup
class and ProxyGroupMember interface work together to maintain a bi-
directional relationship between the two classes.  ProxyGroup is aware
of its members and ProxyGroupMember is aware of which group it is a
member of.

The ProxyGroupMember provides API hooks for the class implementing the
class to be notified when being added or removed from a Proxy Group.
It is intended that the ProxyInterface class to inherit from
ProxyGroupMember.

This change will replace much of the functionality of the GroupManager
class.

After some testing, it was found that the GroupManager class creates
difficulties for adding and removing elements from the group, which
then requires many function calls.  This new ProxyGroup-
ProxyGroupMember pattern addresses the issues that were faced before,
while still providing the ability to effectively unittest the
components.

TEST=unittest
BUG=chromium:863260

Change-Id: Ie1e5a08847de149e58104cea7357cc0b7b3d244d
Reviewed-on: https://chromium-review.googlesource.com/1175552
Commit-Ready: Alex Dale <sigquit@google.com>
Tested-by: Alex Dale <sigquit@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>