shill: Add new init script for sandboxing kill switch
We have a feature flag in Chrome (called ShillSandboxing) that
will control whether shill is run in a minijail sandbox. Add an Upstart
init job that queries Chrome for this policy and writes a file
accordingly that can be accessed upon next boot. This init job will
start on the "started system-services" Upstart event, which assures
that Chrome is running but doesn't require the user/guest to have logged
in.

CQ-DEPEND=CL:1176106,CL:1173260,CL:1204652
BUG=chromium:649417
TEST=disable/enable ShillSandboxing Chrome feature flag (through
/etc/chrome_dev.conf on DUT) and see shill running as root/non-root
Change-Id: I45a665561b8809f8cbef42a09d6cd35884f0bfe7
Reviewed-on: https://chromium-review.googlesource.com/1176166
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>