crash: crash_sender: make random generator more robust
POSIX arithmetic expansion requires support of at least signed long
integers.  On 32-bit systems, that may mean 32-bit signed integers,
in which case the 32-bit random number read from /dev/urandom may be
interpreted as negative when used inside an arithmetic expansion
(since the high bit might be set).  While bash/dash don't appear to
do this, mksh (on Android) at least is known to behave this way, and
we really should be robust in case the shell changes.
http://pubs.opengroup.org/onlinepubs/009695399/utilities/xcu_chap02.html#tag_02_06_04

To handle this, simply grab 3 bytes instead of 4 bytes of random data.
This reduces our entropy slightly, but the only user is the code that
generates a random delay before uploading a crash (with a max value of
600), so it's not a big deal.

Steve Fung noticed this originally, but this is a simpler fix.

BUG=None
TEST=precq passes

Change-Id: I6b908f7451a7d16c59ba074cd6af1e7eee1541fd
Reviewed-on: https://chromium-review.googlesource.com/1179099
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>