portier: Neighbor cache expiry mechanism.
This changes provides the NeighborCache class the ability to discard
any cache entries that are older than a specified time frame.  This is
necessary as RTNetlink may not signal the process on entries being
deleted.

This also allows for the discarding of locally generated cache entries
which are not managed by the kernel.  These entries are generated when
a downstream interface perform neighbor discovery, requiring the ND
Proxy to cache the entry for two-way traffic to be established.  The
Linux kernel will not cache entries for non-solicited neighbor
advertisement if they were not self initiated.

TEST=unittest
BUG=chromium:863260

Change-Id: I3e8839ca0eecb46f5f1a85bc840cf5a8c7138a86
Reviewed-on: https://chromium-review.googlesource.com/1176393
Commit-Ready: Alex Dale <sigquit@google.com>
Tested-by: Alex Dale <sigquit@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>