vm_tools: Add MIME type passing from container to host
This parses the /etc/mime.types file in the container and passes that
information along to the host so that the container's MIME types can be
used for file extension mappings after we use the hardcoded ones in
Chrome OS.  Chrome OS has a list that is not very extensive, so this
helps more file types get recognized properly by their corresponding
app. It will also parse $HOME/.mime.types if present and use that as
overrides to the system settings.

BUG=chromium:869747
TEST=Unit tests pass, Open As shows proper options now
CQ-DEPEND=CL:1188447

Change-Id: I6945cf8abaed97ec4f61fd22d16f02bc820d9f4a
Reviewed-on: https://chromium-review.googlesource.com/1188793
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>