portier: use of sendmsg in EtherSocket instead of sendto.
This change attempts to add a small performance post to EtherSocket by
removing the need to excessively copy data.  After a review of
ICMPSocket (in cl:1167990), it was suggested to use sendmsg() instead
of sendto() system call.  This removes the need for heap allocated
header fields and copying of packet payload in to a temporary buffer.

TEST=none
BUG=chromium:863260

Change-Id: I094d4c6c489f3926086cc10ccffba258da4f0f23
Reviewed-on: https://chromium-review.googlesource.com/1176464
Commit-Ready: Alex Dale <sigquit@google.com>
Tested-by: Alex Dale <sigquit@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>