crash: fix unit tests to use ScopedTempDir everywhere
Some of the unit tests manually create their own temp directories
from a relative path.  This creates a temp folder in the folder
that the test is run from.  This is a problem when using test_droid
or a non-interactive ADB shell command, since the root directory is
not writable.  Instead use a ScopedTempDir to handle temporary
folders.

This is based on Steve's work, but also rebased to the latest tree
and simplified a bit.

BUG=b:26671657
TEST=precq passes

Change-Id: Idcee64814e60c8587e8c6d49ea41008f6adb96c7
Reviewed-on: https://chromium-review.googlesource.com/1176445
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>