crash: anomaly_collector: improve filter logic a bit
Some of the sub-collectors are inconsistent in their handling of the
filter flag.  Push it down into the common AnomalyStart helper so we
can delete it from the individual helpers.

We also allow AnomalyEnd to run, but we use /bin/cat to dump the log
to stdout.

BUG=chromium:874698
TEST=precq passes

Change-Id: I4ff32659b5271b5f1973f177cefcc4f6d6652e8d
Reviewed-on: https://chromium-review.googlesource.com/1214633
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>