crash: kernel_collector: Use constexpr for global constants
This patch changes all the global constants in the anonymous namespace
of kernel_collector.cc from const to constexpr. Because its better
somehow. Except for the ones that construct objects because that can't
be done at compile time, and the ones that are arrays of strings
because... reasons. I don't really understand any of this anyway.

BUG=None
TEST=unit tests

Change-Id: I50777e7ef0db23ffdc73dd423c9902e15e6ae5bf
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1182318
Reviewed-by: Mike Frysinger <vapier@chromium.org>