bluetooth: Add parameter to PauseDiscovery and UnpauseDiscovery
PauseDiscovery and UnpauseDiscovery are going to be used by multiple
clients and not all requests are related to system suspend/resume.
Therefore a parameter is needed for clients to indicate whether the
request is related to system suspend/resume, which in turn affects how
bluetoothd records metrics for chip lost (ignore if adapter is gone
during suspend/resume).

BUG=chromium:874611,chromium:523176
TEST=Tested with the modified bluetoothd
CQ-DEPEND=CL:1184151

Change-Id: I61363c9b80ea49cc32b269f67b20252487014815
Reviewed-on: https://chromium-review.googlesource.com/1184154
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Qiyu Hu <qiyuh@chromium.org>