metrics: memd: initialize tracing and don't collect NR_PAGES_SCANNED
This fixes two bugs revealed by changes in behavior on 4.14 kernel.
Firstly, the code was not initializing tracing_enabled and tracing_on,
which it should have, and it was just luck that it worked on other
platforms.  Secondly, 4.14 no longer outputs nr_pages_scanned in
/proc/vmstat, which is just as well because apparently it was getting
reset in a way that made it not useful.

BUG=chromium:875467
TEST=ran on device

Change-Id: Ie4f4d376822ab2d6e3a2e0139eed800b4fc1b5ea
Reviewed-on: https://chromium-review.googlesource.com/1187190
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>