init: preserve demo mode resources in clobber-state
Updates clobber-state to include the factory-installed Demo Mode
offline resources in the list of files to preserve during a "safe"
powerwash.

These offline resources are installed in the factory flow. They cannot
be changed, but they are deleted once the device is determined to be
in active non-retail use. Until then, we want to preserve them during
powerwashes, for two reasons:

1. The factory flow includes a factory reset which would remove them:
   echo "fast safe keepimg" \
   >/mnt/stateful_partition/factory_install_reset && reboot

2. Devices improperly configured in stores will need to be powerwashed
   to correctly enter demo mode, e.g. with Ctrl+Alt+Shift+R or by
   enabling and disabling Dev Mode

The offline resources are installed as an ImageLoader image. The total
size is expected to be about 500 MiB. This is much larger than the other
files being preserved, but can still fit comfortably in RAM during the
clobber-state.

BUG=chromium:877866
TEST=created these files on the device, and powerwashed using the
     methods above. verified the files were still in place after
     powerwash even if they totaled 1 GiB in size. tested on 3 different
     models. also created other files in the same directory to verify
     they were not preserved.

Change-Id: Ibe533c79d484797bfbdda5a1ca0105bb3dd2171a
Reviewed-on: https://chromium-review.googlesource.com/1188395
Commit-Ready: Michael Giuffrida <michaelpg@chromium.org>
Tested-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: Greg Levin <glevin@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>