init: bypass firmware update error.
boot-update-firmware.conf would call the firmware update script, and
clear the boot message if any of script showed the message. However,
if the script is terminated with error, the boot message won't be
cleared.

This CL bypasses the error. Then we can guarantee the boot message is
cleared no matter the scripts are executed successfully or not.

The verification step:
  1. Install M67 stable test image
  2. Flash local built test image to USB
  3. Boot from USB, then the boot message is shown
With this CL, the boot message is cleared.

BUG=chromium:869155
TEST=Check the boot message is cleared

Change-Id: I66b07e0d7aca2397d002ade2fc25e00f7d5abf5d
Reviewed-on: https://chromium-review.googlesource.com/1195283
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Tai-Hsu Lin <sheckylin@chromium.org>