attestation, cryptohome: The delegate can get the EK iff it has rights
Add a new field to remember whether the TPM delegate has rights to
read data (e.g. the public EK). Because such rights have been granted
to the delegate since M66 but without any record of them, we will try
first if the field isn't set and remember if we failed to read the EK
with a non retryable error as meaning that the delegate can't read
the public EK.

In addition, we update the semantics of the GetEnterpriseEnrollmentId
and ComputeEnterpriseEnrollmentId calls to return false for transient
errors and true for success or definitive errors. This means that
callers of those methods (which are exposed over DBUS by the service)
should only retry when they have a chance to actually get better
results.

BUG=chromium:867724
TEST=unit tests for samus (TPM 1.2) and eve (TPM 2.0)

Change-Id: I913857ffdd382f220cf0fd6a78b914d5cd76ac95
Reviewed-on: https://chromium-review.googlesource.com/1196841
Commit-Ready: Yves Arrouye <drcrash@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>