metrics: memd: mount /dev/chromeos-low-mem in sandbox
This fixes a bug (not filed) that caused memd to almost
never enter fast-poll mode.  In addition to fixing the bug,
this adds a warning message when the underlying condition
occurs.  Furthermore, this changes the condition for
switching from slow-poll to fast-poll to a fixed distance
(difference) between "available" and "margin", rather than
a multiplicative factor.

BUG=none
TEST=tested with platform_SimpleMemoryPressure

Change-Id: If582161faddcf59df45f8d38d80bab4eaabfb2e3
Reviewed-on: https://chromium-review.googlesource.com/1205376
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Cheng-Yu Lee <cylee@chromium.org>