cecservice: set logical address type on dev init
According to the CEC docs, there is a class of TVs that drop EDID while in standby mode.
For such TVs cecservice has a special handling where it sends a 'image view on' message
from unregistered logical address to wake them up, as prescribed by the docs.

However, before this change, cecservice was making setting logical addresses after EDID has
shown up. That is a problem because the CEC framework requires logical addresses to be set
before an attempt to send any messages is made. In case of this type of TVs that results
in request being sent before we attempt to set logical addresses, which makes the framework
reject such request with EPERM error.

This change makes the cecservice to call set logical addresses on device init.

BUG=b:114000740
TEST=Run unit test and also manually checked that attempt to wake up TVs when there is no
  EDID is not rejected by kernel with EPERM error. Unfortunately I don't have an access
  to such TV to fully verify that there are no additional problems with handling of such
  TVs, but at least the obvious bug preventing this from working is fixed.

Change-Id: If691710beff3b6d2dc20b7576b8e69f3d1952e50
Reviewed-on: https://chromium-review.googlesource.com/1206353
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Dariusz Marcinkiewicz <darekm@google.com>
Reviewed-by: Felix Ekblom <felixe@chromium.org>