shill: revise wifi/cellular metrics
Here we collect histograms for daily/monthly online times
on wifi/cellular, as well as fractions of wifi and cellular
over total connection time, limited to samples collected
when both technologies are present and a specific one is
chosen.

These histograms do not distinguish the case of a built-in
cellular modem vs. a USB dongle.  Unfortunately it is not
easy to tell them apart because they both show up as USB
devices.

This adds a total of 10 histograms:
[any / cellular / wifi / cellular fraction / wifi fraction] x
[daily / monthly]

BUG=chromium:832145
TEST=ran on device

Change-Id: I92dd76e2f303ee49e24d499acc1e0738f66ba6a0
Reviewed-on: https://chromium-review.googlesource.com/1237139
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>