attestation: do not insert empty NVRAM quotes into the identity data
The ACA does not expect empty quotes and will reject them as
erroneous, preventing enrollment. We also only have code setting
NVRAM quotes if USE_TPM2 is defined (mainly because the Cr50 NVRAM
constants are defined in a package that is only a dependency for
TPM 2).

This CL also quotes both PCRs and NVRAM indices using a loop, allowing
both sets of quotes to easily be extended without copying and pasting
the same logic over and over as was done until now.

BUG=chromium:882049
TEST=enroll with the ACA on a device without NVRAM data, e.g. eve

Change-Id: I218a1fdfd062215f32a05e427379312d07fb7af7
Reviewed-on: https://chromium-review.googlesource.com/1214643
Commit-Ready: Yves Arrouye <drcrash@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>