shill: Record UMA metric for shill sandboxing
Report Chrome OS metric signifying whether a system is running shill
in a sandbox or not. This is controlled through a Finch flag, but
requires 1-2 reboots to take affect. This metric will help us ensure
that shill is being run with sandboxing enabled/disabled at the
frequency that we expect.

BUG=chromium:649417
TEST=made sure shill.sh doesn't fail, observed the metric being
recorded in /var/log/metrics/uma-events

Change-Id: I580ac9ed47a4d77cbdfdbb075a83dd2ca728e559
Reviewed-on: https://chromium-review.googlesource.com/1216503
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>