init: Only run recover_duts on lab machines
recover_duts is intended for recovery of DUTs in the lab, however it
is started on any device with a test image. Currently the script runs a
single hook ('check_ethernet'), which only performs actual work
on lab devices.

Limit the execution of 'recover_duts' to lab machines. The now redundant
check in 'check_ethernet' is removed in a separate CL.

This also silences the 'audit spam' messages about 'check_ethernet' on
the serial console of developer devices with 'console=ttyXYZ' set in the
kernel command line.

BUG=chromium:881932
TEST=on a device without /mnt/stateful_partition/.labmachine:
    ps -ef | grep recover_duts | grep -v grep
      => no output
  on a device without /mnt/stateful_partition/.labmachine:
    ps -ef | grep recover_duts | grep -v grep
      => script is running
CQ-DEPEND=CL:1217723,CL:1226161

Change-Id: I5da84dcdeb54948e731562ab6dc2e52ec27e06c2
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1217825
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>