libpasswordprovider: Change keyring permission to allow the user access to the password.
The user's login password is saved to the keyring by session_manager,
which is run as root. Shill can be run as root or shill depending on
the feature flag enabled. This is meant to be a temporary measure, the
goal is to eventually run shill as the shill user only.

To support the scenario of running shill as root, this change changes
the password keyring permission to allow the root user to read it.

This change should be removed once the shill sandboxing flag is
removed (tracked in http://crbug.com/875028).

BUG=chromium:649417
TEST=run unit tests on cave, deploy libpasswordprovider to cave and
test that shill running as root and shill can access the password.

Change-Id: I76e5dae743af160641ff31c6a0a66daaf86f9482
Reviewed-on: https://chromium-review.googlesource.com/1217827
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: May Lippert <maybelle@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>