login: Make Subprocess class more testable.
Some functionality in the Subprocess class really belonged in
SystemUtils. Besides being more consistent, now this functionality can
be mocked, and the Subprocess class unit tested.

This is in preparation for adding functionality to session_manager to
isolate (parts of) the user session.

BUG=chromium:216789
TEST=Unit tests pass, guest login works, regular login works.

Change-Id: I6d58b421feecd8f1332b55cbdcffc7e5cc483bd1
Reviewed-on: https://chromium-review.googlesource.com/1220127
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>