cros_component: show warnings upon upload
Once a new component release is uploaded to gs bucket, Omaha picks it up
and never scan it again. So any modification to the release after upload
does not have effect.

We show a warning before the upload operation and prompt user to input
'Yes' to actuall upload to the repo.

BUG=chromium:883357
TEST=manual testing

Change-Id: Idf888ef31c6b11080d242cf21b687061f9b2046e
Reviewed-on: https://chromium-review.googlesource.com/1222526
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Xiaochu Liu <xiaochu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>