smbprovider: Remove invalid tests
This change removes a couple of tests that do not make sense in a real
world scenario. These tests were verifying that non-files/non-dirs were
not returned via ReadDirectory, but this does not really make sense
because the type returned from a ReadDir+ (which is now all ReadDirs) is
libsmb_file_info which represents a File or Directory.

BUG=chromium:757625
TEST=unittests
Change-Id: I9369ecf9e8873d25c59b60c080254f02799b244b
Reviewed-on: https://chromium-review.googlesource.com/1224635
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Reviewed-by: jimmy gong <jimmyxgong@chromium.org>