debugd: Allow specifying the encoding of logs
This change adds a way to specify the encoding of logs instead of
guessing. That way logs that are supposed to be text can have Unicode
replacement characters inserted (instead of being base64-encoded), and
binaries can be base64-encoded from the get-go to avoid paying the price
of having to scan the whole thing for an invalid character.

BUG=b:29988671
TEST=FEATURES=test emerge-${BOARD} debugd

Change-Id: I4a1054958da5c8fe7f266b3aeecc1103a6855c91
Reviewed-on: https://chromium-review.googlesource.com/1226214
Commit-Ready: Luis Hector Chavez <lhchavez@chromium.org>
Tested-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>