installer: Fix another 4k-sector issue in copy_partition
The copy_partition function had a condition that was both comparing
source and destination block counts as if they were the same, and
assuming 512-byte block sizes. Both parts of these condition do not
work if one of the devices has a block size that is not 512 bytes.

BUG=chromium:832160
TEST=Run chromeos-install on Cheza, observe sda12 now gets copied.

Change-Id: I8d7855c8b610edf831e151869206f1d7648041d0
Reviewed-on: https://chromium-review.googlesource.com/1228606
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>