diagnostics: Constant for Mojo pipe token
Add a constant with the Mojo pipe token for the Mojo pipe
between the browser and the diagnosticsd daemon. This constant
needs to be known on both ends of this Mojo pipe.

For example, on the browser side, that's the name that is passed
to mojo::OutgoingInvitation::AttachMessagePipe().

BUG=chromium:869377
TEST=system_api package builds

Change-Id: I3020c07de5493dac0740e43a4ec2d1791105e805
Reviewed-on: https://chromium-review.googlesource.com/1230873
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>