metrics: fix memd mmap seccomp filter
Before the fix, the filter was applied to the length argument.
It worked because the length was always a multiple of a
sufficiently large power of 2.  Of course it would not have
blocked the calls we intended to block.

Note that the first three arguments of mmap, mmap2, and
mprotect have the same meaning: addr, len, prot.
The mprotect filter was already correct.

Thanks to Andrew for noticing this :)

BUG=none
TEST=none

Change-Id: I651a2d782872aa26de0ba260f5ce782dc2157e80
Reviewed-on: https://chromium-review.googlesource.com/1234019
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>