intel: ipu3: Fix 1080p still capture distorted issue
The configuration from the app is as below:
one 3264x2448 YUV output stream,
one 1080p BLOB output stream,
one 1080p IMPL output stream.

The IMPL output stream runs in video pipe.
The YUV output stream and BLOB output stream
all run in still pipe. Because the YUV steam is the bigger,
the still pipe will out 3264x2448. And the BLOB's
source buffer is the 3264x2448, but the width:height ratio
is not the same as the out resolution 1080p.
The crop function is not implemented, so the issue happens.

Crop the 3264x2448 to 3264x1836, then let it to be the source
buffer of the JPEG encoder. The JPEG encoder will do the
downscaling and JPEG encoding to 1080p.

BUG=b:116070581
TEST=Build and deploy package to platform and camera works

Change-Id: I4df4dc9b11a5b674c15fcb0ae34553f4a0fb7741
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1234714
Commit-Ready: Wei Meng <wei.meng@intel.corp-partner.google.com>
Tested-by: Wei Meng <wei.meng@intel.corp-partner.google.com>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Liang L Yang <liang.l.yang@intel.com>