vm_tools: sommelier: Add demo app flags for setting width/height
This also adds a bgcolor flag and defaults the value to what we expect
in the tests in order to maintain backwards compatibility. The width and
the height arguments are in screen pixels. For wayland we reduce that by
the scale factor, which wayland then scales up by the scale factor. This
then results in both x11 and wayland demo apps giving the same size
window if each is given the same width/height parameters.

BUG=chromium:849438
TEST=Manually verified on eve

Change-Id: Ia770b7d4c63128f65402a50c621f785eff3b0abc
Reviewed-on: https://chromium-review.googlesource.com/1308692
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>