vm_tools: concierge returns status from StartVm
A VM started with |start_termina| set to |true| is not
usable to start lxd containers until the Tremplin process
has started for that VM.

Concierge now listens for the TremplinStartedSignal from
cicerone and saves a bool internally.

StartVmResponse now returns a status for the VM based on this bool.

Once this lands, Chrome's CrostiniManager can use the status to
avoid waiting for the TremplinStartedSignal. This signal can
be missed if a vm is started via vmc in crosh.

BUG=b:884576
TEST=Manual testing on eve

Change-Id: I2a6697444b3ceb4a4acece1bd15eaec707eab26f
Reviewed-on: https://chromium-review.googlesource.com/1240215
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Josh Pratt <jopra@chromium.org>