login: Allow powerwash for enterprise devices
We're in the process of building a new remote command (remote wipe/
powerwash), and the idea is to re-use the existing StartDeviceWipe DBus
call.  The problem is this is currently blocked for enterprise/managed
devices, and that is precisely the set of devices we want to target.  This
check will be removed from CrOS login_manager, and bubbled up into Chrome
code which does the call currently (so the new code will have the check
skipped).

Here is the DD for the feature go/remote-powerwash-command

BUG=chromium:827523
TEST=None.

Change-Id: Icd2424a034a9d97e2ee855141018a280013fc337
Reviewed-on: https://chromium-review.googlesource.com/1236053
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ivan Šandrk <isandrk@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Ivan Šandrk <isandrk@chromium.org>