diagnostics: Correct upstart job respawn limit
Put reasonable respawn limit into the upstart jobs for the diagnosticsd
and the diagnostics_processor daemons.

The default respawn limit, which is 10 times within a 5 second period,
is too liberal for our daemons, since it takes about a second to
kill-and-restart such daemon on an average Chromebook.

BUG=chromium:869377
TEST=manual (start a buggy version of the daemon and ensure it
     doesn't get restarted infinitely)

Change-Id: I57528f31fdbf5fcfda4e054528369ae011b0c880
Reviewed-on: https://chromium-review.googlesource.com/1238513
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>