shill: stick with pre-v4.6 rtnl_link_stats64
If we upgrade our linux-headers, we pick up the new, larger
rtnl_link_stats64, and shill starts failing to handle link statistics:

  [WARNING:device_info.cc(1258)] Link statistics size is too small: 184 < 192

The problem is that upstream Linux commit 6e7333d315a7 ("net: add
rx_nohandler stat counter") added a new field to this struct. We don't
need to handle it, and rtnl otherwise handles variable-length messages
just fine. We just need to not complain if we see the 184-byte version
(pre 4.6) or the 192-byte version (post 4.6).

This will avoid copying the rx_nohandler field.

BUG=none
TEST=with upgraded linux-headers,
     `/usr/local/lib/flimflam/test/list-devices | grep ByteCount`
     shows reasonable results

Change-Id: Ic65ba440945ad4e37d1b37d35036f71d60fbae91
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1241421
Reviewed-by: Ben Chan <benchan@chromium.org>