vm_tools: garcon: Add independent thread for D-Bus communication
This fixes the threading model in garcon to avoid deadlocks relating to
communication with cicerone. We add a 3rd thread which is for D-Bus
communciation with PackageKit. This prevents contention with blocking
operations from the host over gRPC which required use of the D-Bus
thread (that was also being used for file system watching, and could
then cause invocations back to the host, resulting in outbound
communication from the host timing out).

BUG=chromium:888745
TEST=tast run vm.CrostiniStartEverything, manually verified as well

Change-Id: I24cd72cf541c86355a46c15eee9df54f8e3ed18d
Reviewed-on: https://chromium-review.googlesource.com/1244167
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>