init: use minijail static mode for rsyslogd.
Preload mode mixes minijail enforcement into part of the executed
process, which makes hard for SELinux to describe the behavior of the
program without allowing the domain to mount/chroot/etc.

Patching libminijailpreload to setcon seems infeasible since /proc is
highly likely to be mounted as readonly in the minijail.

BUG=b:116072767
TEST=boot and observe dmesg

Change-Id: Ibfd43c8add504ef783e6ac95dfbc0c4d3a64636e
Reviewed-on: https://chromium-review.googlesource.com/1248324
Commit-Ready: Qijiang Fan <fqj@google.com>
Tested-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>