shill: openvpn: escape username and tag sent to OpenVPN
According to the OpenVPN documentation
(openvpn/doc/management-notes.txt), the username and password passed to
OpenVPN through the OpenVPN Management interface can be in quotes where
special characters like double quote and backslash must be escaped with
with a leading backslash. This CL fixes shill to properly escape the
username as well as the tag associated with the username/password
command passed to OpenVPN. Practically speaking, the tag associated with
the username/password command is extracted from a prior notification
from OpenVPN, which shouldn't contain special characters that need
escaping. But to be extra paranoid, the tag is escaped as well.

BUG=chromium:886911
TEST=Run unit tests.
TEST=Run network_VPNConnect.openvpn* tests.

Change-Id: I7c9cd715c110cf7cba16316c9d3a2f78fc348952
Reviewed-on: https://chromium-review.googlesource.com/1250131
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>