vm_tools: Add uninstall functionality to garcon
Add the ability to uninstall Linux packages to garcon. Note that the
package-to-be-uninstalled is identified by the .desktop file instead of
by the package ID. We send the package_id along with the .desktop file in
UpdateApplicationList; however, nothing watches for package upgrades, so
the package_id may be stale by the time the uninstall happens. (The
package_id in UpdateApplicationList is still useful for deciding IF a
.desktop file can be auto-uninstalled, so I'm leaving it in place.)

BUG=chromium:822514
TEST=Uninstalled Eclipse using cicerone_client. Tried multiple
simultaneous uninstalls; checked that one failed with
other-uninstall-is-in-progress message. Tried various combinations of
simultaneous installs and uninstalls, ensured only one worked at a time.

Change-Id: I5e112226c4e22e868151851580ad05bef72eb938
Reviewed-on: https://chromium-review.googlesource.com/1249673
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>