intel: ipu3: Fix video and still pipe outputs unsynchronized issue
For the video pipe and still pipe, if the n input frame
is sent to the IMGU, the output frame is
coming from the (n - 1) input frame.
For the first input frame (n = 0) for both pipes,
the output frame is dummy frame.

The still pipe is not running all the time,
it will be asked to run in some request.
To skip the dummy frame,
the still pipe will be fed
with one same input frame for two times.

And the video pipe is working continuously.

So to make sure the two pipes outputs are the same,
the input frame for the still pipe must be
the last one which is fed to video pipe.

BUG=b:115849890
TEST=Build and deploy package to platform and camera works

Change-Id: I47a12e772d7f453b1c28e700af31879301fb4423
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1250409
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Liang L Yang <liang.l.yang@intel.com>