vm_tools: cicerone: Handle async operations in the client
This updates the client so that the CreateLxdContainer and
InstallLinuxPackage commands will not return until the D-Bus signal has
been returned indicating completion. Previously they returned upon
successfully starting those operations to let them complete in the
background. The plan is to update crosh to manage containers which is
why this is being done.

BUG=None
TEST=Verified container creation & package install manually

Change-Id: If473bfe25fff91cede664358f8ab16d46b8348b9
Reviewed-on: https://chromium-review.googlesource.com/1252541
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>